--- Coding/CreditCard/CreditCard.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public interface CreditCard {
-
-
-    void attach(User u) throws invalidPreferenceException;
-}
