--- Coding/CreditCard/Notification.java
@@ -1,8 +1,0 @@
-package org.example;
-
-public interface Notification {
-
-    void notify(User u);
-
-    void publicNotify();
-}
