--- Coding/CreditCard/invalidPreferenceException.java
@@ -1,8 +1,0 @@
-package org.example;
-
-public class invalidPreferenceException extends Exception{
-
-        public invalidPreferenceException(String error){
-            super(error);
-        }
-}
