+++ Coding/hw2/Q1/Q1.java
