+++ Coding/hw2/Q1/Shape.java
@@ -1,0 +1,25 @@
+package org.example;
+
+public class Shape {
+    //polymorphism
+
+    public double getPerimeter(){
+        return 0;
+    }
+
+}
+
+class Circle extends Shape {
+
+    private double r;
+
+    public Circle(double r){
+        this.r = r;
+    }
+
+    public double getPerimeter(){
+        double v = 3.14 * 2 * r;
+        return v;
+    }
+
+}
