+++ Coding/hw3/CreditCard/CreditCard.java
@@ -1,0 +1,7 @@
+package org.example;
+
+public interface CreditCard {
+
+
+    void attach(User u) throws invalidPreferenceException;
+}
