+++ Coding/hw3/CreditCard/Notification.java
@@ -1,0 +1,8 @@
+package org.example;
+
+public interface Notification {
+
+    void notify(User u);
+
+    void publicNotify();
+}
