+++ Coding/hw3/CreditCard/invalidPreferenceException.java
@@ -1,0 +1,8 @@
+package org.example;
+
+public class invalidPreferenceException extends Exception{
+
+        public invalidPreferenceException(String error){
+            super(error);
+        }
+}
