+++ Coding/hw3/hw3-collections/ArraysExerciseTest.java
