+++ Coding/hw3/hw3-collections/HashSetExerciseTest.java
