+++ Coding/hw5/Q2/MyThreadExtends.java
@@ -1,0 +1,9 @@
+package org.example;
+
+public class MyThreadExtends extends Thread{
+    @Override
+    public void run(){
+        System.out.println("New thread extends thread");
+        }
+
+}
