+++ Coding/hw5/Q2/MyThreadRunnable.java
@@ -1,0 +1,9 @@
+package org.example;
+
+public class MyThreadRunnable implements Runnable{
+
+    @Override
+    public void run() {
+        System.out.println("New Thread using Runnable");
+    }
+}
