+++ Coding/hw5/TwoThreadA.java
@@ -1,0 +1,31 @@
+package org.example;
+
+public class TwoThreadA {
+    private static final Object monitor = new Object();
+    private static int value = 1;
+
+    public static void main(String[] args) {
+        PrintRunnable runnable = new PrintRunnable();
+        new Thread(runnable).start();
+        new Thread(runnable).start();
+    }
+
+    static class PrintRunnable implements Runnable{
+        @Override
+        public void run(){
+            synchronized (monitor){
+                while(value <= 10){
+                    System.out.println(Thread.currentThread().getName() + ": " + value++);
+                    monitor.notify();
+                    try{
+                        if (value < 11){
+                            monitor.wait();
+                        }
+                    } catch (InterruptedException e){
+                        e.printStackTrace();
+                    }
+                }
+            }
+        }
+    }
+}
